refactor: BackGround 로직 분리 & Service에서 sleep 메서드 호출
